Refactored retread/[id]
